pyproject.toml
[tool.poetry]
name = "statement-sensei"
packages = [
    {include = "webapp"}
]
version = "0.3.5"
description = "PDF to CSV conversion for your bank statements"
authors = ["Benjamin Dornel <benjamindornel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
streamlit = "^1.36.0"
monopoly-core = "0.9.5"
pybadges = {git = "https://github.com/benjamin-awd/pybadges"}


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"
pylint = "^3.1.0"



[tool.poetry.group.build.dependencies]
git-cliff = "^2.3.0"
pyinstaller = "^6.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.pylint]
max-line-length=120
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "logging-fstring-interpolation"
]
